Initial Spec
tl;dr: A procedure to encode English/Latin characters as base-10
integers, and vice versa!

Initial draft specification of the kidbase10 encoding protocol. The
procedure is inspired by fuller encoding protcols such as ASCII, but
simplified so it resembles a substitution cipher. The process/goal is
hopefully intuitive for a child who can understand base-10 numbers (up
to 1 billion in value) and has moderate proficiency in written English
or other language which uses Latin characters.

This protocol is meant to be an instructional tool to teach
programming and computer science concepts. It is not meant to
faithfully transmit or encrypt data - DO NOT USE IN PRODUCTION!

Signed-off-by: Adam B Kaplan <adam@adambkaplan.com>