spec: Allow Alternate Character Sets and Forms
Allow variants of the spec to use alternate character encodings. This
is meant to encourage use of non-Latin characters and other
combinations of characters for encoding input/output. The current set
of Latin characters is drawn from common use in English - this may not
apply to other languages which use Latin characters, and certainly will
not work for languages with non-Latin character sets.

Allow alternate forms of characters in the encoding specification. This
would allow use of uppercase and lowercase Latin characters to be used
interchangeably as input to be encoded. To ensure consistency, outputs
from decoding MUST choose a canonical character form (ex: always print
uppercase) in these situations. The specification uses Latin uppercase
and lowercase as an example, however with non-Latin characters this is
not the only instance of "alternate character forms" in language
character sets. Ex: "end" forms of Hebrew and Arabic characters.

Signed-off-by: Adam B Kaplan <adam@adambkaplan.com>